roofer-app: add reconstruct
